Fix top layer of pressure-latitude plots (#3)
Our existing pressure-latitude climate change plots have a row of NaNs in the top atmosphere layer.  It happens this is because the pressure at the layer midpoint inferred from ACE output is greater than the pressure at the layer midpoint with the calculated "standard" pressure levels assuming a surface pressure of 1000 hPa, and our pressure interpolation routine does not extrapolate.  It is sort of strange that this occurs, because this should be a pure pressure layer, meaning its midpoint should not depend at all on the surface pressure.  

The source of this issue is that when computing the standard pressure thicknesses, we used:
```
standard_delp = (ak + 100_000 * bk).diff("pfull")
```
which results in a `float64` dtype array, but the pressure thicknesses we inferred from ACE using something equivalent of:
```
ace_delp = (ak + ps * bk).diff("pfull")
```
results in a `float32` array.  If we ensure that `standard_delp` is a `float32` array by casting `100_000` to a `float32` value, the pressure at the top layer midpoint in the standard levels and ACE end up identical, and therefore no extrapolation—or interpolation for that matter—is required.  We make the same update in the `noise_floor.py` code to ensure this change is reflected in the metrics as well.

This does not substantially change the qualitative story, but it is nice to be able to clearly see the stratospheric cooling.