Use 100 bins for histogram instead of 50 (#8)
This gives a higher fidelity view of the histogram.

Previously we normalized the percent change by the change in global mean temperature in the reference case, but here we normalize by the model-specific global mean temperature change to be more accurate. It doesn't make a huge difference, but it is cleaner.